# swagger.yml
openapi: 3.1.0
info:
  title: Receipt Wrangler API.
  termsOfService:
    there are no TOS at this moment, use at your own risk we take no
    responsibility
  license:
    name: GPLV3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 5.0.0
servers:
  - url: /api
security:
  - bearerAuth: [ ]
paths:
...
  /systemEmail/{id}:
    get:
      tags:
        - SystemEmail
      summary: Get system email by id
      description: This will get a system email by id
      operationId: getSystemEmailById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Id of system email to get
      responses:
        200:
          description: The system email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemEmail"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
    put:
      tags:
        - SystemEmail
      summary: Update system email by id
      description: This will update a system email by id
      operationId: updateSystemEmailById
      requestBody:
        description: System email to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSystemEmailCommand"
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Id of system email to update
        - in: query
          name: updatePassword
          schema:
            type: boolean
          required: true
          description: Whether or not to update the password
      responses:
        200:
          description: The updated system email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemEmail"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
    delete:
      tags:
        - SystemEmail
      summary: Delete system email by id
      description: This will delete a system email by id
      operationId: deleteSystemEmailById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Id of system email to delete
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
  /systemEmail/getSystemEmails:
    post:
      tags:
        - SystemEmail
      summary: Gets paged system emails
      description: This will return paged and sorted system emails
      operationId: getPagedSystemEmails
      requestBody:
        description: Paging and sorting data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagedRequestCommand"
      responses:
        200:
          description: Paged system emails
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedData"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
      security:
        - bearerAuth: [ ]
  /systemTask/getPagedSystemTasks:
    post:
      tags:
        - SystemTask
      summary: Gets paged system tasks
      description: This will return paged system tasks
      operationId: getPagedSystemTasks
      requestBody:
        description: Paging and sorting data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetSystemTaskCommand"
      responses:
        200:
          description: Paged system tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedData"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
      security:
        - bearerAuth: [ ]
...
  /systemSettings:
    get:
      tags:
        - SystemSettings
      summary: Get system settings
      description: This will get system settings
      operationId: getSystemSettings
      responses:
        200:
          description: The system settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
        500:
          $ref: "#/components/responses/Internal"
        403:
          $ref: "#/components/responses/Forbidden"
    put:
      tags:
        - SystemSettings
      summary: Update system settings
      description: This will update system settings
      operationId: updateSystemSettings
      requestBody:
        description: System settings to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSystemSettingsCommand"
      responses:
        200:
          description: The updated system settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
        500:
          $ref: "#/components/responses/Internal"
        403:
          $ref: "#/components/responses/Forbidden"
  /systemSettings/restartTaskServer:
    post:
      tags:
        - SystemSettings
      summary: Restart task server
      description: This will restart the task server
      operationId: restartTaskServer
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
        403:
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Request was successfully processed
    BadRequest:
      description: The request was malformed
    Forbidden:
      description: The request was not allowed
    Internal:
      description: There was an error processing the request
  schemas:
    ReceiptStatus:
      description: Status of a receipt
      type: string
      enum:
        - "OPEN"
        - "NEEDS_ATTENTION"
        - "RESOLVED"
        - "DRAFT"
        - ""
...
    GroupStatus:
      type: string
      enum:
        - "ACTIVE"
        - "ARCHIVED"
...
    SortDirection:
      type: string
      enum:
        - asc
        - desc
        - ""
...
    OcrEngine:
      type: string
      enum:
        - "TESSERACT"
        - "EASY_OCR"
    SystemTaskStatus:
      type: string
      enum:
        - "SUCCEEDED"
        - "FAILED"
...
    AssociatedEntityType:
      type: string
      enum:
        - "NOOP_ENTITY_TYPE"
        - "RECEIPT"
        - "SYSTEM_EMAIL"
        - "RECEIPT_PROCESSING_SETTINGS"
        - "PROMPT"
...
    CurrencySeparator:
      type: string
      enum:
        - ","
        - "."
    CurrencySymbolPosition:
      type: string
      enum:
        - "START"
        - "END"
...
    BaseModel:
      type: object
      required:
        - id
        - createdAt
      properties:
        id:
          type: integer
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
          default: 0
        createdByString:
          type: string
          description: Created by entity's name
          default: ""
        updatedAt:
          type: string
          default: ""
    Category:
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the category
        description:
          type: string
          description: Description of the category
        updatedAt:
          type: string
          x-go-name: UpdatedAt
      description: Category to relate receipts to
...
    Comment:
      required:
        - comment
        - receiptId
        - userId
        - id
      type: object
      properties:
        additionalInfo:
          type: string
          description: Additional information about the comment
        comment:
          type: string
          description: Comment itself
        commentId:
          type: integer
          description: Comment foreign key used for repleis
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        receiptId:
          type: integer
          description: Receipt foreign key
        updatedAt:
          type: string
        userId:
          type: integer
          description: User foreign key
          format: uint64
      description: User comment left on receipts
    FileData:
      required:
        - receiptId
        - id
      type: object
      properties:
        createdAt:
          type: string

          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
          x-go-name: CreatedBy
        fileType:
          type: string
          description: MIME file type
          x-go-name: FileType
        id:
          type: integer
          format: uint64
          x-go-name: ID
        imageData:
          type: array
          description: Image data
          items:
            type: integer
            format: uint8
          x-go-name: ImageData
        name:
          type: string
          description: File name
          x-go-name: Name
        receiptId:
          type: integer
          description: Receipt foreign key
          format: uint64
          x-go-name: ReceiptId
        size:
          type: integer
          description: File size
          format: uint64
          x-go-name: Size
        updatedAt:
          type: string

          x-go-name: UpdatedAt
      description: File data for images on a receipt
      x-go-package: receipt-wrangler/api/internal/models
    Group:
      required:
        - name
        - status
        - id
        - groupMembers
        - isAllGroup
        - groupReceiptSettings
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        groupSettings:
          $ref: "#/components/schemas/GroupSettings"
        groupReceiptSettings:
          $ref: "#/components/schemas/GroupReceiptSettings"
        groupMembers:
          type: array
          description: Members of the group
          items:
            $ref: "#/components/schemas/GroupMember"
        id:
          type: integer
          format: uint64
        isDefault:
          type: boolean
          description: Is default group (not used yet)
        name:
          type: string
          description: Name of the group
        isAllGroup:
          type: boolean
          description: Is all group for user
        status:
          type: string
          $ref: "#/components/schemas/GroupStatus"
        updatedAt:
          type: string
      description: Group in the system
...
    GroupMember:
      required:
        - groupId
        - groupRole
        - userId
      type: object
      properties:
        createdAt:
          type: string
        groupId:
          type: integer
          description: Group compound primary key
          format: uint64
        groupRole:
          type: string
          $ref: "#/components/schemas/GroupRole"
        updatedAt:
          type: string
          x-go-name: UpdatedAt
        userId:
          type: integer
          description: User compound primary key
          format: uint64
          x-go-name: UserID
      description: Group member
...
    Receipt:
      required:
        - amount
        - date
        - groupId
        - name
        - paidByUserId
        - status
        - id
      type: object
      properties:
        amount:
          type: string
          description: Receipt total amount
        categories:
          type: array
          description: Categories associated to receipt
          items:
            $ref: "#/components/schemas/Category"
        comments:
          type: array
          description: Comments associated to receipt
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        date:
          type: string
          description: Receipt date
        groupId:
          type: integer
          description: Group foreign key
          format: uint64
        id:
          type: integer
          format: uint64
        imageFiles:
          type: array
          description: Files associated to receipt
          items:
            $ref: "#/components/schemas/FileData"
        name:
          type: string
          description: Receipt name
        paidByUserId:
          type: integer
          description: User paid foreign key
          format: uint64
        receiptItems:
          type: array
          description: Items associated to receipt
          items:
            $ref: "#/components/schemas/Item"
        resolvedDate:
          type: string
          description: Date resolved
        status:
          $ref: "#/components/schemas/ReceiptStatus"
        tags:
          type: array
          description: Tags associated to receipt
          items:
            $ref: "#/components/schemas/Tag"
        updatedAt:
          type: string
        createdByString:
          type: string
          description: Created by string, which is anything that is not a user
      description: Receipt
    Tag:
      required:
        - name
      type: object
      properties:
        createdAt:
          type: string
          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Tag name
        description:
          type: string
          description: Tag description
        updatedAt:
          type: string
      description: Tag to relate receipts to
...
    TagView:
      required:
        - id
        - name
        - numberOfReceipts
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the tag
        description:
          type: string
          description: Description of the tag
        updatedAt:
          type: string
        numberOfReceipts:
          type: integer
          description: Number of receipts associated with this tag
      description: Tag to relate receipts to
...
    PagedRequestCommand:
      required:
        - page
        - pageSize
      type: object
      properties:
        page:
          type: integer
          description: Page number
        pageSize:
          type: integer
          description: Number of records per page
        orderBy:
          type: string
          description: field to order on
        sortDirection:
          type: string
          $ref: "#/components/schemas/SortDirection"
...
    PagedData:
      required:
        - data
        - totalCount
      type: object
      properties:
        data:
          type: array
          items:
            anyOf:
              - $ref: "#/components/schemas/Receipt"
              - $ref: "#/components/schemas/Category"
              - $ref: "#/components/schemas/Tag"
              - $ref: "#/components/schemas/Prompt"
              - $ref: "#/components/schemas/Group"
              - $ref: "#/components/schemas/TagView"
              - $ref: "#/components/schemas/SystemTask"
              - $ref: "#/components/schemas/ReceiptProcessingSettings"
              - $ref: "#/components/schemas/SystemEmail"
        totalCount:
          type: integer
...
    SystemEmail:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            host:
              type: string
              description: IMAP host
            port:
              type: string
              description: IMAP port
            username:
              type: string
              description: IMAP username
            password:
              type: string
              description: IMAP password
    UpsertSystemEmailCommand:
      type: object
      required:
        - host
        - port
        - username
        - password
      properties:
        host:
          type: string
          description: IMAP host
        port:
          type: string
          description: IMAP port
        username:
          type: string
          description: IMAP username
        password:
          type: string
          description: IMAP password
    TaskQueueConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            name:
              $ref: "#/components/schemas/QueueName"
            priority:
              type: integer
              description: Queue priority
    UpsertTaskQueueConfiguration:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/QueueName"
        priority:
          type: integer
          description: Queue priority
    SystemSettings:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - taskQueueConfigurations
          properties:
            enableLocalSignUp:
              type: boolean
              description: Whether local sign up is enabled
              default: false
            currencyDisplay:
              type: string
              description: Currency display
              default: "$"
            currencyThousandthsSeparator:
              $ref: "#/components/schemas/CurrencySeparator"
            currencyDecimalSeparator:
              $ref: "#/components/schemas/CurrencySeparator"
            currencySymbolPosition:
              $ref: "#/components/schemas/CurrencySymbolPosition"
            currencyHideDecimalPlaces:
              type: boolean
              description: Whether to hide decimal places
              default: false
            debugOcr:
              type: boolean
              description: Debug OCR
              default: false
            numWorkers:
              type: integer
              description: Number of workers to use
              default: 1
            emailPollingInterval:
              type: integer
              description: Email polling interval
              default: 1800
            receiptProcessingSettingsId:
              type: integer
              description: Receipt processing settings foreign key
            fallbackReceiptProcessingSettingsId:
              type: integer
              description: Fallback receipt processing settings foreign key
            taskConcurrency:
              type: integer
              description: Concurrency for task worker
              default: 10
            taskQueueConfigurations:
              type: array
              items:
                $ref: "#/components/schemas/TaskQueueConfiguration"
    UpsertSystemSettingsCommand:
      type: object
      required:
        - currencyThousandthsSeparator
        - currencyDecimalSeparator
        - currencyHideDecimalPlaces
        - currencySymbolPosition
        - taskConcurrency
      properties:
        enableLocalSignUp:
          type: boolean
          description: Whether local sign up is enabled
        currencyDisplay:
          type: string
          description: Currency display
        currencyThousandthsSeparator:
          $ref: "#/components/schemas/CurrencySeparator"
        currencyDecimalSeparator:
          $ref: "#/components/schemas/CurrencySeparator"
        currencySymbolPosition:
          $ref: "#/components/schemas/CurrencySymbolPosition"
        currencyHideDecimalPlaces:
          type: boolean
          description: Whether to hide decimal places
        debugOcr:
          type: boolean
        numWorkers:
          type: integer
          description: Number of workers to use
          default: 1
        emailPollingInterval:
          type: integer
          description: Email polling interval
        receiptProcessingSettingsId:
          type: integer
          description: Receipt processing settings foreign key
        fallbackReceiptProcessingSettingsId:
          type: integer
          description: Fallback receipt processing settings foreign key
        taskConcurrency:
          type: integer
          description: Concurrency for task worker
        taskQueueConfigurations:
          type: array
          items:
            $ref: "#/components/schemas/UpsertTaskQueueConfiguration"
...
    SystemTask:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            type:
              $ref: "#/components/schemas/SystemTaskType"
            status:
              $ref: "#/components/schemas/SystemTaskStatus"
            startedAt:
              type: string
            endedAt:
              type: string
            associatedEntityId:
              type: integer
            associatedEntityType:
              $ref: "#/components/schemas/AssociatedEntityType"
            ranByUserId:
              type: integer
              format: uint64
            resultDescription:
              type: string
            childSystemTasks:
              type: array
              items:
                $ref: "#/components/schemas/SystemTask"
    GetSystemTaskCommand:
      allOf:
        - $ref: "#/components/schemas/PagedRequestCommand"
        - type: object
      properties:
        associatedEntityId:
          type: integer
          description: Associated entity id
        associatedEntityType:
          $ref: "#/components/schemas/AssociatedEntityType"
    ReceiptProcessingSettings:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            name:
              type: string
              description: Name of the settings
            description:
              type: string
              description: Description of the settings
            aiType:
              $ref: "#/components/schemas/AiType"
            url:
              type: string
              description: URL for custom endpoints
            key:
              type: string
              description: Key for endpoints that require authentication
            model:
              type: string
              description: LLM model
            isVisionModel:
              type: boolean
              description: Is vision model
            ocrEngine:
              $ref: "#/components/schemas/OcrEngine"
            prompt:
              $ref: "#/components/schemas/Prompt"
            promptId:
              type: integer
              description: Prompt foreign key
...
    Prompt:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - name
            - prompt
          properties:
            name:
              type: string
              description: Prompt name
            description:
              type: string
              description: Prompt description
            prompt:
              type: string
              description: Prompt text
...